Adding entry for i18n/L10n to the Roadmap
I skimmed through the project roadmap and think that it can certainly use a brief introduction to internationalization (i18n) and localization (L10n) as well as good common practices to make web projects localizable and globalized. I'm going to familiarize myself with the details of what has been provided so far in other parts and try to come up with some content for this section.