Node.js Additional resource
I found this simple Meteor beginner tutorial a great way to get into Node.js development. It should also be a lot easier to grasp for students that have completed the Ruby on Rails section, because Node.js tutorials are more like vanilla Ruby tutorials whereas this Meteor tutorial is a lot more similar (although a little smaller) than the Ruby on Rails tutorial from the Ruby on Rails section.