Changed third link and text on JS Basics.
The current link to JQuery's Javascript 101 returns a 404 error, so it would be nice to remove that link and add a new one or simply remove that 3rd part if it's not necessary. 

Here, I substituted the Mozilla JS Basics page, which a lot of people direct beginners to. It covers things a little differently than the original link, but provides a good review of the Codecademy course while showing some of the cool ways that you can use JS on the web. As a beginner, I've been itching to see some basic use of JS on a webpage and this site scratched that itch.