DELETE is a specialed version of POST not GET
In the tutsplus  HTTP tutorial explains that PUT and DELETE are specialized versions of a POST request not GEt, since POST is used to create a resource, it could be used to do a deletion or an update. Here is the section where he explains that:
```
* **GET**: fetch an existing resource. The URL contains all the necessary information the server needs to locate and return the resource.
* **POST**: create a new resource. POST requests usually carry a payload that specifies the data for the new resource.
* **PUT**: update an existing resource. The payload may contain the updated data for the resource.
* **DELETE**: delete an existing resource.

The above four verbs are the most popular, and most tools and frameworks explicitly expose these request verbs. PUT and DELETE are sometimes considered specialized versions of the POST verb, and they may be packaged as POST requests with the payload containing the exact action: create, update or delete.
```