add James Kelsey's solution
I'm really proud of this solution, so I would like to add it here. I used as much Ajax as I could and played with modals. The log in with Facebook only works for me because Facebook changed their system recently and won't allow me to register my app into production mode because it recreates the functionality of Facebook.