Refactor lesson completion model spec file
Addresses Issue 262

I replaced the two FactoryGirl variables with doubles, added shoulda matchers to test the validations and associations, and removed the last test. I did not address the existing comment about testing that the lesson completion is deleted when a lesson or user is deleted—slightly beyond my scope of knowledge.