Adjusted images sizes and corrected quote
Although the screen-shots along the lesson are large they are mostly empty black spaces, it's hard to read the text on them (I guess it may depend on the user's screen resolution as well), so I edited them zooming on the text so it could become more readable. Maybe the text turned out too large, but I can adjust it if necessary.
I've also corrected the quote to match the output shown in the image on the topic **Pushing your finished work to github**, #2.