Update project_final_js.md
I'm asking to add as my final Odin project something that departs a little from the stated instructions. It's not a remake of my favorite site, but is a web app I built on Rails and ReactJS for scheduling lessons, so it definitely incorporates the skills I've learned from completing (98%) the Odin Project. It was mostly a personal project, so I thought it may qualify. I'd like to finish the entire Odin curriculum but not sure I have time to do a site remake at this point. But I understand if this doesn't qualify.:)