Fix Failing Tests
I've noticed that a few extra tests are failing that were caused by recenet submissions.
Because there already a lot of tests failing it was easy to miss these extra ones for
the person submitting the PR and the people reviewing the PR.

This fixes most of the failing tests. I've commented out the project controller tests as I intend
to submit a follow PR to refector that spec.

All tests should now be green in the redesign branch, and will make it easier for us to catch any
tests that are broken by a PR.