Allow Users to Sign in with any omniauth provider
This allows a user to sign up/in with both omniauth providers if they
have the same email address associated with them. Previously we solved
this problem with clunky buttons in the users settings page which
gave the user the ability to link omniauth accounts to thier Odin account.
This solution should be much more user friendly.

There are a couple of edge cases that still exist, it would be good to make
people familar with them.

* If the user signs up with an omniauth account first, a random password will be
generated for them. There will be no way to tell what this password is. So if a user
wants to sign in with an email and password with the same account they will first have
to request a password reset.

* If the user signs in with omniauth accounts that have two different emails associated with them. A seperate
Odin account will be created for each one.

These edge cases will probably not happen that often so they are comprimises we are willing to tolerate.