Fill in README
I've created a basic README, it just has the following sectuons so far: introduction, contributing and maintainers.

I figured we needed at least the introduction and contributing section to help any contributors that might want to help
out.

If you would like to add to the README please do, this iteration is very basic and to the point.