Optimize `CoursesController#index` (#890)
Currently, a significant amount of time is spent in loading the lesson
objects. This is because they have a big `content` field, so loading
them is very expensive:

1. The database has to read a lot into memory
2. The database has to send back a big result
3. The Ruby driver has to parse/process the big payload
4. The `content` field need to be loaded/copied into big Ruby strings
5. Having these big strings in memory increases memory usage/GC pressure

Therefore, it is a good idea to avoid loading the lesson objects unless
they are absolutely needed to display the content, which is probably
only needed in `LessonsController#show`.

This patch demonstrates a way to refactor the relevant code to avoid
needing to load the lessons object. The focus here is
`CoursesController#index`, but a lot of other endpoints can benefit
from this optimization (you can find them on Skylight).