Use Current User instead of Finding the User in Courses Controller (#892)
Thanks to great work done this PR https://github.com/TheOdinProject/theodinproject/pull/890
we no longer need to find the current user and assign it to an instance variable set the courses controller. We can
instead use current_user directly and avoid performing a redundant query to find the user.