Replace TryRuby Course with RubyMonk
Codeschool has recently been shut down so we need to replace their TryRuby Course.

I've included the Ruby Monk assignment as the last assignment due to it getting into objects in Ruby. This might be a bit advanced for a student at this level. We should keep an eye out for what students say about it on chat for the next few weeks.