Change Git reading assignments
I changed the Git reading assignments away from Atlassian to just use the ProGit book, which is more thorough but still easy to read (and was already part of the assignment anyway). It also has the added benefit of not discussing SVN. Furthermore, I discovered that at least one of the Atlassian sections was just copy and pasted from ProGit, so we might as well go straight to the source and give them the credit.