Change subheading of "Take Breaks" and edit body
For the section "Pitfalls to Avoid", the subheading "Take Breaks!" did not flow well with the outline of the article.  Example:  Pitfalls to avoid; Procrastination, "Take Breaks!", Digital Distractions, Rabbit Holes, Comparing Yourself to Others.  "Not Taking Breaks" is a much better fit as a subheading for this section.

Edited body paragraph of "Not Taking Breaks" for better readability.