Removes reference to deployed heroku app
I've noticed that many learners are encountering an error because they aren't running the `heroku run rails db:migrate` command. The see the link and immediately navigate to the app. I'm suggesting we remove the statements that invite them to see the link and have them proceed to running `heroku run rails db:migrate`.