Update DOM-manipulation.md
The original para reads as follows:

"See DOM Enlightenment’s section on CSS Style rules for more info on inline styles."

This sentence becomes problematic as it does not refer to where this section is in the book. There is only one chapter and title in the book which contains the term "CSS Style" (Chapter 9 - CSS Style Sheets & CSS rules) however this is not the correct chapter that the course material refers to. 

The correct chapter is in fact Chapter 6 - Element Node Inline Styles.

The change involves:
1. Stating which chapter is to be used, in case the book's formatting changes;
2. Changing the hyperlink to an anchor link for Chapter 6.