Add explanation that addEventListener uses callbacks
I was super confused when I saw that function(e) just worked somehow, and wondered where e was passed in. I got it explained to me on discord that e was passed in by addEventListener because it was defined to try to pass in arguments to the second input. Just seeing this without explanation made me confused, so I think something along these lines should be added.