Update problem_solving.md
Trying to resolve issue #18528 (Pseudo code in ruby problem solving lesson is formatted poorly)

The 2 sections of pseudo code here both render correctly in the Git preview with appropriate line breaks, but on the actual TOP page they do not show any line breaks. I replaced the use of 3 backticks with 3 tildes instead to see if that will fix it. If that doesn't work I could also try using 4 spaces to indent each line or use 2 carriage returns to force the line break.