Proposal to change the assignment
Hello,

I watched the both videos. The content is the same, but the video I propose is much shorter (because the information is condensed and presented in a scripted manner) and more helpful. The end of the video contains a practice section (which back then would net you an award - sticker), it isn't relevant but it provides a visual representation of how real git usage would look like. 

Proposed video: https://youtu.be/HkdAHXoRtos
Current video: https://www.youtube.com/watch?v=HVsySz-h9r4

Matt,