JavaScript does have real actual classes
ECMAScript editors have explicitly said JS classes are real. Brian Terlson, for one example, was/is spec editor from ES7-present (you'll find his name in the spec for those editions). He occasionally comments on reddit, including [this comment](https://www.reddit.com/r/javascript/comments/4bmiqj/using_classes_in_javascript_es6_best_practice/d1avktu/). "It is not at all helpful for beginners to constantly be exposed to the notion that JS classes are 'fake'," he says, "They are not."

JS classes actually behave the same as classes in Python, Ruby, or Smalltalk, among others. If you write `instance.foo` in Python, for example, then it searches "instance" for a property named "foo". If it doesn't find it, then it follows a special link to another object and searches that other object for "foo". If it doesn't find it, then it follows that object's special link to another object, and so on until it finds a "foo" property or until the chain of object links ends. Here's a look at [JavaScript and Python classes side-by-side](https://i.imgur.com/p9Kw815.png).