Update oop-concepts.md
I'm proposing these changes to solve Issue #21006 

I have clarified some of the wording on the Single Responsibility Principle. Then after the code blocks I've added in the other description of the principle (as a single reason to change) per the linked articles.

I added in the curriculum format for Assignment and Additional Resources sections and then moved the articles down into the appropriate categories. I also included the actual article titles in the linked text, since I always find this more helpful in staying organized with the assignments.

Thanks for considering these updates!

Daniel Murphy