Update conditional_logic.md
Although the code at issue is clearly just an example demonstrating a `case` statement, I do not think it should contain code that is syntactically incorrect. Here, the example shows that an integer is allowed to contain commas. This is not true, and would result in the following error if attempted:

```ruby
> foo = 0
> foo += 50,000

SyntaxError ((irb):53: syntax error, unexpected ',', expecting end-of-input)
foo += 50,000
         ^
```