Re-arranged Article positions in "assignments"
the first article was about why Class is good or bad and was really confusing without actually understand what class is. I think it would be best moved to the 3rd position instead of 1st. So the first 2 readings explain what class is and how its used. and then the last reading talks about varying opinions. Some people doing this course don't have any knowledge of other languages so  this is our first exposure to Class (so the article on why its good/bad really was just confusing).