Swapping embarrassing for complimentary
In an effort to be more inclusive swapping embarrassing complimentary
created an internet we all want to be a part of. I've been told that
this is a joke and I'm reading too much in it, but to error on the side
of being nice I think is better than to error on the side of creating
cultures that harrass people anonymously.

Swapping "embarrassing" for "complimentary" is an easy switch, that is
able to keep the same context for the purpose of learning in the project.