Removes references to Windows, WSL
This page feels like it should focus on the actual installations. We discourage Windows in the lesson prior. I'm also going to expand the language there so it feels repetitive to address it again here.

Also made a minor change about what OS we recommend. Added remarks about also support official flavors of Ubuntu but not distros that are based on Ubuntu.