Add screen reader recomendation
One thing that really made me able to commit to starting this project was downloading a screen reader. I was not very used to reading articles like this before that, and found the fact that most resources are articles to be very discouraging. I think a screen reader recommendation could help lots of people who are in the same boat