React-router documentation for github pages
I encountered a pretty frustrating problem while deploying it to gh-pages. most of it was due to github not detecting the routes i've added. I tried searching on the internet regarding the solution and after 1-2 hours of trial and error came across this
https://github.com/facebook/create-react-app/issues/1765
solution by react team to somehow use tricks to bypass this problem.
I wonder if people here got into the same problem while deploying router based react page. I suggest Odin team to add some steps or instructions to better deploy router based projects on the assignment page such that noone else has to suffer the same. 
However, @todoroki in discord room suggested to create a PR for the same.