Update introduction_to_the_backend_lesson.md
I found this lesson quite lacking, in terms of both content and grammar, and I had to do quite a bit of Googling to gain even a basic understanding of what the back end actually is.

I deleted the second resource in the Assignment, as I would argue that it's more a promo piece about getting into back-end development/being a back-end developer than the back end itself. I added two resources - the best I could find on the subject, after viewing quite a lot of search results.

I added basic definitions of the front end and back end, which I feel are required, and made a number of grammatical revisions.

Finally, I don't really understand the second learning outcome question. Perhaps it's the wording. What reason IS there to install back-end languages on one's computer other than to be able to run them...? The reasons do not appear to be "listed above"...