Updated install to include bundler gem
In the next lesson (basic data types) bundler is required in order to properly use the practice exercises. While it may seem intuitive for some of us to install it, not having it installed already may confuse some users, so I have added an instruction for users to do it here.