Standardizes language.
The idea is that using "front-end" or "back-end" as an adjective to something like "languages" (ex. "back-end languages") is grammatically appropriate. When referring to the backend or frontend as nouns - ideas within full-stack web development - that we join the words together. The consensus is that this will be our standard in language going forward.