Add Additional Resources Description
Someone had messaged me on TOP Facebook saying the DOM session had a ton of links and was really long and asked if they had to do it all. After looking it over, I noticed that the Additional Resources section didn't have the line saying these are optional. Since it is one of the earlier lessons I think it is important to have that in there. It was probably unintentionally missing from this lesson.