Making context of sharing screenshots clearer.
Changing it to clarify that if your sharing a screenshot that you should also share the respective code in a repository. Rather than just solely having a screenshot when asking for help. Might make it clearer that if there is correlating code to share that in a repository rather than just a screenshot?