Fix: code tag color inside knowledge check
`<code>` tag inside `.knowledge-check-link` which is an `<a>` tag, overrides the color of `<code>` to `grey` from `pink`. It's because bootstrap's reboot which has a selector `a > code` that changes it. So I've added a `.knowledge-check-link > code` tag which gives `pink` color to `<code>` tag. `$pink` variable is declared in bootstrap.