Remove handlebars from Express-Lesson-1.md
Please note this is my first time ever contributing to an open-source project, so apologies in advance if I haven't proposed this correctly! Below I have copied my feedback which I previously posted on the TOP 'contributions-suggestions-bugs-discussions'  discord channel. One of the core admins (BriggsE) suggested I make the pull request.

The lesson notes included a mention that students should consider using handlebars or EJS rather than the pug (previously Jade) template framework while completing the MDN tutorial. I followed this advice, and used handlebars.

I really wish I hadn't, as relying on handlebars has opened multiple cans of worms and detracted from focusing on the core parts of the lesson (i.e learning how to use express). For one thing, handlebars has made a major change to fix a security vulnerability (read more here: http://mahmoudsec.blogspot.com/2019/04/handlebars-template-injection-and-rce.html) which has in turn disrupted the flow of the MDN tutorial (which relies heavily on virtual properties on mongoose models, which handlebars no longer allows).

In addition to this, I have found that handlebars lacks basic key functionality / documentation. For example, something as simple as including an "else if" statement doesn't seem to be possible, while pug seems to make that easy. I also had to discover through trial and error how template extensions work with handlebars in express, as I couldn't find any documentation on this topic.

All in all I've wasted quite a bit of time dealing with handlebars and I think TOP could be majorly improved if more resources were provided to introduce the framework before attempting to use it in the MDN project. And given that security issue, it might be better if students just stayed away from it entirely. I ended up getting around that security block with this package, but the author made it clear in the notes that this is NOT a good solution and should not be used in the "real world". (read more: https://www.npmjs.com/package/@handlebars/allow-prototype-access).

CHANGES MADE:

- Removed references to handlebars in the Templating Engines section.
- Added a hyperlink to the first pjs mention to like directly to the ems getting started page.