Update head element section in html-boilerplate.md
There is a statement "Anything included within the head element will not be displayed to the user." which may be misinterpreted by beginners that "whatever you write inside <head> and </head> tags even the plain text should not be displayed on webpage by browser." which is wrong. I think the statement might create this misconception which will remain in their mind until they try stuff like putting random text in head tag and then they can see the output on screen.