Slight typo in comments.
Hey, while going through the different methods when I saw comments explaining the methods,
I got a bit confused with the comment  //select the container div. 

Like does the command selects the div or are we required to select the div ourselves after that command.
I think it should be selects as was in the 4th comment.
I have made similar changes in other comments as well.