Update oop-concepts.md
This contains the copy-editing changes made in a previous pull request. It was requested that I split my changes into two requests.

It is not necessary to continue to quote or bold terms that have already been explained. It is also not necessary to put quote marks around capitalised proper nouns. All this creates inconsistency.