fixed broken image in css_math_functions.md
The image was linking to an album instead of the actual file. The image was badly cropped and not very informative so I created a new image that looks better and gets the point across better. The previous anchor link was also pointing to the imgur album, which is useless, so I changed it to point towards the codepen being discussed.