Update how_does_the_web_work.md
The "Additional Resources" section of the page has been duplicated. The original section was left in place above "Knowledge Check" and an additional section was added underneath.

It seems that the standard is to have the "Additional Resources" section at the bottom of the page. I propose to remove the duplicated section and to move the original section to the bottom where the duplicated section was.