Update block-and-inline.md
The link behind "like this" was taking to a song video i.e. Rick Astley - Never Gonna Give You Up. I think it would be more helpful if there was an actual ideo that is explaining "Block and Inline elements". So I have proposed an update that is relevant for the learners. The link will redirect to a short youtube video, exaplaining "Block and Inline elements" in detail. I hope this would be more helpful for upcoming learners.

note: I'm not the owner of the video. I just searched youtube and found this video which seems pretty straight forward and informative.

I hope this change will be approved for the benefit of learners.