Update to first example of specificity
In the first example for specificity, changed .subsection to .list, so the point on precedence makes more sense. As .subsection, there is no conflict, so precedence simply doesn't apply, even if Rule 2 is more specific. Changing .subsection to .list, will actually show how the more-specific rule takes precedence in what would otherwise be a conflict.