Proposed copy refreshment for recipe project
The changes I've proposed are obviously not do or die, but I think some of the copy throughout the Foundations curriculum could be refreshed/cleaned up a bit and I'm happy to go through and help polish some of that copy.

While the Odin model is to force developers to do the work themselves, I think implementing more example code and more full explanations of the assignments at hand would quicken a learner's experience as they go through the curriculum. Open to any and all feedback!