target="_blank"
Adding this attribute to the knowledge check section, in my opinion, boosts the user experience slightly since someone may want to read the linked resource in a new tab and not the same tab that has the Odin Project. I'd recommend implementing this change in other sections where there are anchor tags as well.