Change adjective for the game instructions
There is no need to refer to the Wikihow article on how to play Rock Paper Scissors as ridiculous. It cannot be assumed that everyone knows how to play it, and thus it is not correct to mock those who are not familiar with the game.
ri·dic·u·lous - adjective
deserving or inviting derision or mockery; absurd.

The new adjective (detailed) not only avoids this problematic behavior, but is also objective instead of subjective.