Add Knowledge Check: Fundamentals Part 5
Let's get those learner's brains ticking! I was hit with this problem during an interview and, despite being familiar with the Fibonacci algo, I struggled working through the steps to realizing the pattern (still solved it, but took way too long).  If learners are completing Fibonacci in Fundamentals Part 4, I think this would be an excellent next step.