redirect certain links path to a new tab
on line 240, this link redirects you to the webpage but on the same tab of the course. which is annoying since most users like to be redirected to a new tab for accessibility purposes. TRY adding -->target="_blank" rel="noopener noreferrer"<-- this inside the link tag as an attribute next to the page link.