Update project_binary_search_trees.md
I think what's being asked for in these method tasks is a little confusing. See the below excerpt from the discord. Now that I've seen this, I understand, but I can definitely understand Melvin's confusion. Maybe an example or just a mention of the nodes being modified via the function in some way would help clarify? It's a bit hard to parse as is.

melvin — 09/26/2022
Binary Search Trees Task 7 Write inorder, preorder, and postorder functions that accept a function parameter. Each of these functions should traverse the tree in their respective depth-first order and yield each node to the provided function given as an argument. The functions should return an array of values if no function is given.
What does  The functions should return an array of values if no function is given. mean? The functions should take an input and if no input is provided they should return an array of values? Where do the values come from?

nevz — 09/26/2022
The values are each node you've visited during the traversal, so each time you've visited a node, the value of that node will be pushed to an array. 

If you've passed a function to those methods for example levelorder((el) => el * 2)) then each value will be multiplied by 2. Otherwise, you will just return the array without any modification.

You should be returning something like:
[21, 10, 156, 4] // levelorder IF no function
[42, 20, 312, 8] // levelorder IF there is a function (el) => el * 2