Feature: New Rails Deployment Lesson
Because:
* Heroku no longer has a free tier.

This commit:
* Remove Heroku install step from Installing Ruby lesson - it was never needed there.
* Remove Rails "Preparing for Deployment" lesson - we no longer need it.
* Change Rails mailing lessons to refer to generic "email providers" instead of the Heroku Sendgrid add-on.
* Remove references to Heroku from the Rails final project.
* Change Rails Deployment lesson to...
  * Allow users to choose from a range of hosting providers instead of just Heroku.
  * Further explain what a hosting providers are, what a PaaS provider is and how they work.
  * More tips for troubleshooting and debugging deployment issues.

Additional Notes:
The rationale for recommending a range of options instead of just one is because there is no like for like alternative for Heroku's free tier. All providers have a very limited free tier.

We will be advising using the free tiers of a combination of different hosting providers for free hosting. We also mention the option is there to choose one hosting provider and pay for it if the learner has the means.

Our learners come from a lot of different backgrounds and don't all have the means to pay for hosting. The different circumstances of each of our learners must be given the up-most respect by providing everyone with options so we can help as many people as possible.