Change font family name in example
The generic font family fallback (sans-serif) is now consistent with
the specified font family name.

An alternative would have been to keep the previous name ("Times New
Roman") and change the generic font family to be
consistent (serif). This solution has the advantage of showing
separators in complex names for both the font family name (which uses
spaces, and therefore needs to include quotes in this instance) as
well as for the generic font family (which uses dashes, but that
detail would be lacking if we switched to merely serif).