Rewrite small parts of Events intro
Removed explanation of method 1 that you can "you can set the 'on_event_' property ". 
This is not an actual property and googling it reveals only people who've also followed 
the Odin project and copied this explanation verbatim. Replaced with explanation of 
generic property name and gave a concrete example by mentioning 'onclick', which is 
used in the examples of Method 1 and Method 2 below this introductory text.

Also replaced usage of "DOM objects" with "DOM nodes", which is consistent with 
the rest of the lesson and makes it more obvious that Method 2 and Method 3 are 
working with exactly the same JavaScript objects.