Ruby Project: Event Manager: Rewrite Initial Setup opening paragraph (#25155)
* Reword Initial Setup to match previous curriculum

In the previous TOP Curriculum, throughout Foundations, it is assumed that all projects are cloned into your repos directory. Since the Ruby and JS paths are meant to take place after Foundations, it is assumed that you will follow previous practices. This means cloning the repo and knowing the basic terminology and commands. 

I don't recall if mkdir or touch was covered, but the cloning should be second nature at this point, and I think knowing what CLI stands for would also be a part of that knowledge.

* Update ruby/files_and_serialization/project_event_manager.md

Co-authored-by: Kevin Mulhern <kevinmulhern23@gmail.com>

* Update ruby/files_and_serialization/project_event_manager.md

Co-authored-by: Kevin Mulhern <kevinmulhern23@gmail.com>

---------

Co-authored-by: Kevin Mulhern <kevinmulhern23@gmail.com>