overhaul: Changing the Git Commit Message Editor
Git Basics - Changing the Git Commit Message Editor

https://www.theodinproject.com/lessons/foundations-git-basics#changing-the-git-commit-message-editor

I think this section is great & important.  But I think we can probably improve the wording of it.
I found it confusing the first couple of times reading it.

I've noticed others have found it confusing too, 
when they've gone to commit without -m and couldn't easily figure out how to fix this (without asking for help), 
because they've moved past it without entering the command correctly.

I've drafted a modification to the lesson's text for this section.