Update time_complexity.md
I noticed that this sentence " Why is Big Omega and why isn’t it as useful? ’in this lesson is using incorrect grammar. I updated it to " What is Big Omega and why isn’t it as useful?’. I checked the lesson with the MD preview tool. Rendered as expected.