Update fundamentals-2.md
The MDN section linked, leads to "String Concatenation".

Concatenation is the term for joining 2 strings together.
Using it for embedding variables/expressions is valid but should be irrelevant to the terminology (because coercion takes place, which means we are actually working with strings). 
The correct terminology used for embedding variables/expressions is string interpolation.

The proposed change, is the minimal change making the context of the knowledge check correct.