Added Linux beginner course playlist from @ApoorvGoyalMain
Hello,

I have added a new resource to the Linux documentation section. This is a beginner-friendly course playlist created by @ApoorvGoyalMain on YouTube, which I believe will be useful for new learners. The playlist covers the basics of Linux, including the command-line interface, file management, and system administration.

I have added a link to the playlist in the "Linux Resources" section of the documentation, along with a brief description of what the course covers. I have also included the creator's name and a link to their YouTube channel for reference.

Please let me know if there are any issues with the changes I have proposed.

Thank you,

Zainab Nisa