Add example for rendered lazy numbering in lists
Lazy numbering specifically might be unintuitive for someone new to
Markdown. Especially if a list isn't very long, it could be mistaken for
a typo and therefore modified by someone not directly working on
conforming a lesson to the Layout Style Guidelines as such.
Presenting the rendered result can improve clarity.