Correct grammar mistake (#25426)
A grammar mistake was found in this lesson. Under the jQuery header, in
the first paragraph, you can find the following sentence that I
corrected: "One of the biggest reasons it's begun to fall out of favor
is that you simply don't need it anymore." The phrase "it's begun",
should be corrected to "it has begun". Here is a list of two english
verb conjugation websites proving this:

1. https://conjugator.reverso.net/conjugation-english-verb-begin.html

2. https://cooljugator.com/en/begin