Update project_shopping_cart.md
Changed the order in which TOP students are informed about <HashRouter /> and clarified how it should be implemented within their source code. It was confusing to me initially as I thought <HashRouter /> was only for my deployed product (npm run deploy), but by implementing the instructions earlier with a more detailed description, I believe it will aid future students in not being confused in that sense.